+++ src/SH1/assembler.cpp
@@ -105,7 +105,7 @@ void assembler::assemble(std::string& data)
 				{
 					if (utils::string::is_number(data[1]))
 					{
-						switchnum = std::stoul(data[1]);
+						switchnum = static_cast<std::uint16_t>(std::stoul(data[1]));
 						inst->size += 7 * switchnum;
 					}
 					else
@@ -130,7 +130,7 @@ void assembler::assemble(std::string& data)
 	}
 
 	LOG_DEBUG("assembly file parse complete.");
-	LOG_DEBUG("%lu functions staged for assemble.", functions.size());
+	LOG_DEBUG("%zu functions staged for assemble.", functions.size());
 
 	this->assemble(functions);
 }
